refactor : 코치 피드백 반영  - 기존 API 엔드포인트는 RESTful하지 않음. "유저"에 대한 잔액 조회이기 때문에 balance가 아닌 user로 변경
