+++ src/main/java/com/hanghae/ecommerce/interfaces/user/UserDto.java
@@ -1,0 +1,4 @@
+package com.hanghae.ecommerce.interfaces.user;
+
+public class UserDto {
+}
